components/web-api/requirements.txt
h5py
cython
flask
fasttext
mecab-python3
